Added horizontal scroll hinting and Chrome mouse scroll modifications
To give the user a hint that they are supposed to scroll horizontally,
rather than the regular vertical orientation, we now scroll the photos
across the page at a rate of 20px/s.

For Google Chrome users we can go a stage further and remap their mouse
scroll wheel so that vertical scroll actually moves the page
horizontally. This may be compatible with other Webkit based browsers,
but some more testing is required before it's safe to release into the
wild.